This is the first Halo SCRIPT
This script runs and plays the Halo Theme with selected rhythms and no musical rests or rubato. The idea is to take the sections of the code which can be functionalized and do so, while leaving in the stuff that stands as parameters or write a GUI which facilitates the creation of note and rhythm arrays.