Add first commit of the Simulation
This version of the Simulation has a demonstration of maneuvering thrusters being used to change the orientation of a cylindrical object. There are known bugs, but I am going to merge this into Master anyway as a starting point, and open a new branch on which to fix those bugs.