renamed and add file
Upper directory name has been changed to suggest that all directories and files beneath it are written in MATLAB and require a MATLAB application to run. This will be important when we start making separate directory trees for code in different languages and to serve different purposes. Just keep everything segregated... :0

Also added a small class that controls point mass kinematics.